+++ R/ColumnLinkedMatrix.R
@@ -136,14 +136,16 @@ rownames.ColumnLinkedMatrix <- function(x) {
 
 # This function looks like an S3 method, but isn't one.
 colnames.ColumnLinkedMatrix <- function(x) {
-    names <- NULL
-    if (!is.null(colnames(x[[1L]]))) {
-        p <- dim(x)[2L]
-        names <- rep("", p)
-        nodes <- nodes(x)
-        for (i in 1L:nrow(nodes)) {
-            names[(nodes[i, 2L]:nodes[i, 3L])] <- colnames(x[[i]])
+    nodes <- nodes(x)
+    names <- rep("", nodes[nrow(nodes), 3L])
+    for (i in seq_len(nrow(nodes))) {
+        nodeNames <- colnames(x[[i]])
+        if (!is.null(nodeNames)) {
+            names[(nodes[i, 2L]:nodes[i, 3L])] <- nodeNames
         }
+    }
+    if (all(names == "")) {
+        names <- NULL
     }
     return(names)
 }
@@ -204,7 +206,7 @@ dimnames.ColumnLinkedMatrix <- function(x) {
 cbind.ColumnLinkedMatrix <- function(..., deparse.level = 0L) {
     dotdotdot <- list(...)
     nodes <- list()
-    for (i in seq_len(length(dotdotdot))) {
+    for (i in seq_along(dotdotdot)) {
         node <- dotdotdot[[i]]
         if (is(node, "LinkedMatrix")) {
             # Extract nodes from LinkedMatrix object
@@ -351,9 +353,8 @@ setMethod("initialize", signature(.Object = "ColumnLinkedMatrix"), function(.Obj
         }
         # Warn if rownames of matrices do not match
         names <- lapply(nodes, rownames)
-        names <- names[!sapply(names, is.null)]
         if (length(names) > 1L && !all(duplicated(names) | duplicated(names, fromLast = TRUE))) {
-            warning("row names of matrix-like objects do not match")
+            warning("row names of matrix-like objects do not match: rownames() only uses the row names of the first node")
         }
     }
     .Object <- callNextMethod(.Object, nodes)