+++ NEWS.md
@@ -1,3 +1,13 @@
+# LinkedMatrix 1.3.1
+
+* Fix `rownames()` (for `RowLinkedMatrix`) or `colnames()` (for
+  `ColumnLinkedMatrix`) returning `NULL` if first node does not have dimnames,
+  but other nodes do. Missing entries will be denoted empty string, just like in
+  base `dimnames()`.
+* Do not exclude NULL when checking if rownames (for `ColumnLinkedMatrix`) or
+  colnames (for `RowLinkedMatrix`) do not match.
+
+
 # LinkedMatrix 1.3.0
 
 * Add [crochet](https://cran.r-project.org/package=crochet) subsetting and
@@ -13,12 +23,14 @@
 * New `sort` parameter in `index()` which is set by default.
 * Add examples.
 
+
 # LinkedMatrix 1.2.0
 
 * Fix wrong order in result when subsetting using unordered positive integers.
 * Add `i` and `j` parameters in `index` to only generate entries for those
   indexes.
 * Various subsetting optimizations.
+
 
 # LinkedMatrix 1.1.0
 
@@ -32,6 +44,7 @@
 * Only allow matrix-like objects with matching row or column dimensions (same
   number of rows in case of `ColumnLinkedMatrix` and same number of columns in
   case of `RowLinkedMatrix`) when creating linked matrices.
+
 
 # LinkedMatrix 1.0.0